Minor_2
Minor Update:
- fixed the issue with Encryption Key being unable to decrypt for use to decrypt password - avoided this issue by just not encrypting the session_pw_key because it is almost impossible to decrypt by referencing a decrypted version of itself, maybe need to experiment with setting a constant key for the session_pw_key encryption so it doesnt get regenerated and would be easier to accomplish while still encrypting the database text

+++ SimplePasswordManager.py
@@ -11,17 +11,16 @@ from datetime import date
 
 today = date.today()
 
-#Generate the Fernet encryption key
+# Generate the Fernet Encryption Key
 # Required for encryption and decryption with Fernet as per documentation
-# Since the Encryption Key is generated upon each session,
-# The session's key is converted into a STRING from BITS
-# Then the string is encrypted and written to the SQLite database
-# For each password created or added
-# We can then decrypt the Encryption key saved with that record to 
+# Since the Encryption Key is generated upon each session, the session's
+# Key is converted into a STRING from BITS then the string is written to
+# the SQLite database for each password created or added
+# We can then later use the Encryption key saved with that record to 
 # Decrypt and decode the password
 key = Fernet.generate_key()
 fernet = Fernet(key)
-session_pw_key = fernet.encrypt((key.decode()).encode())
+session_pw_key = key
 
 # Connet to database file included in Simple Program Manager
 sql_connection = sqlite3.connect('SPMdb.db')
@@ -60,6 +59,7 @@ another = f" another password?\n{yes_no}"
 select = ["Select a password to ", " by typing the index of the account: "]
 doing = "ing password..."
 done = ["ed password for ", "...\n"]
+empty_list = ["There are no passwords to ", "...\nReturn to Start Menu?\n"]
 mode = ""
 
 
@@ -90,26 +90,35 @@ def add_pw():
     new_username = str(input("Username: "))
     new_pw = str(input("Password: "))
     
-    print(mode + doing)
-    # User inputted value for new_pw is encoded then saved to a new variable per documentation
-    encoded_pw = fernet.encrypt(new_pw.encode())
-    # The variable needs var.decode() when adding to the encrypted passwords table
-    # This converts the values datatype from BITS to STRING
-    # Otherwise it saves to the list as b'var' instead of 'var'
-    # Decode is different to Decrypt, remember to read the docs more
-    # The encoded pw is BITS datatype once encrypted and needs it's own variable
-    sql_cursor.execute(f"""
-                       INSERT INTO passwords 
-                       VALUES('{new_acct}','{new_username}','{encoded_pw.decode()}','{session_pw_key.decode()}','{today}');
-                       """)
-    sql_connection.commit()
+    if new_acct != "" and new_username != "" and new_pw != "":
+        print(mode + doing)
+        # User inputted value for new_pw is encoded then saved to a new variable per documentation
+        encoded_pw = fernet.encrypt(new_pw.encode())
+        # The variable needs var.decode() when adding to the encrypted passwords table
+        # This converts the values datatype from BITS to STRING
+        # Otherwise it saves to the list as b'var' instead of 'var'
+        # Decode is different to Decrypt, remember to read the docs more
+        # The encoded pw is BITS datatype once encrypted and needs it's own variable
+        sql_cursor.execute(f"""
+                        INSERT INTO passwords 
+                        VALUES(
+                           '{new_acct}',
+                           '{new_username}',
+                           '{encoded_pw.decode()}',
+                           '{session_pw_key.decode()}',
+                           '{today}');
+                        """)
+        sql_connection.commit()
 
-    # Return to Start Menu or repeat
-    again = str(input( line + mode + done[0] + new_acct + done[1] + mode + another ))
-    if again.lower() == "y":
-        add_pw()
+        # Return to Start Menu or repeat
+        again = str(input( line + mode + done[0] + new_acct + done[1] + mode + another ))
+        if again.lower() == "y":
+            add_pw()
+        else:
+            start()
     else:
-        start()
+        print("All fields are required")
+        add_pw()
 
 
 # Edit a password from the encrypted passwords table based on it's index
@@ -125,41 +134,50 @@ def edit_pw():
                        """)
     list_pw = sql_cursor.fetchall()
 
-    # Loop through the contents of list_pw in human-readable format
-    place = 1
-    for i in list_pw:
-        print(place, " | ", i[0], " | ", i[-3], " | ", i[-1])
-        place += 1
+    # Check if list_pw is empty to avoid app crash when input expects user to give index of record that exists
+    if list_pw != []:
+        # Loop through the contents of list_pw in human-readable format
+        place = 1
+        for i in list_pw:
+            print(place, " | ", i[0], " | ", i[-3], " | ", i[-1])
+            place += 1
 
-    # User chooses a record that corresponds with record row when ORDER BY account ASC
-    index = int(int(input( line + select[0] + mode.lower() + select[1] )) - 1)
+        # User chooses a record that corresponds with record row when ORDER BY account ASC
+        index = int(int(input( line + select[0] + mode.lower() + select[1] )) - 1)
 
-    # Before updating the password we need to save the returned account and username for the update statement
-    # Ran into errors when using list_pw[index][1 or 0] directly in the SQLite update statement
-    acct = str(list_pw[index][0])
-    usr = str(list_pw[index][1])
-    
-    replace_pw = str(input("New Password: "))
-    print(mode + doing)
-    # User input value for replace_pw is encoded then saved to a new variable
-    replace_encoded_pw = fernet.encrypt(replace_pw.encode())
-    # Remember the new variable needs to be DECODED before adding to the encrypted passwords table
-    # This converts the values datatype from BITS to STRING
-    sql_cursor.execute(f"""
-                       UPDATE passwords 
-                       SET password = '{replace_encoded_pw.decode()}', 
-                       last_modified = '{today}' 
-                       WHERE account = '{acct}' 
-                       AND username = '{usr}';
-                       """)
-    sql_connection.commit()
+        # Before updating the password we need to save the returned account and username for the update statement
+        # Ran into errors when using list_pw[index][1 or 0] directly in the SQLite update statement
+        acct = str(list_pw[index][0])
+        usr = str(list_pw[index][1])
+        
+        replace_pw = str(input("New Password: "))
+        print(mode + doing)
+        # User input value for replace_pw is encoded then saved to a new variable
+        replace_encoded_pw = fernet.encrypt(replace_pw.encode())
+        # Remember the new variable needs to be DECODED before adding to the encrypted passwords table
+        # This converts the values datatype from BITS to STRING
+        sql_cursor.execute(f"""
+                        UPDATE passwords 
+                        SET password = '{replace_encoded_pw.decode()}', 
+                        pw_key = '{session_pw_key.decode()}',
+                        last_modified = '{today}' 
+                        WHERE account = '{acct}' 
+                        AND username = '{usr}';
+                        """)
+        sql_connection.commit()
 
-    # Return to Start Menu or repeat
-    again = str(input( line + mode + done[0] + list_pw[index][0] + done[1] + mode + another ))
-    if again.lower() == "y":
-        edit_pw()
+        # Return to Start Menu or repeat
+        again = str(input( line + mode + done[0] + list_pw[index][0] + done[1] + mode + another ))
+        if again.lower() == "y":
+            edit_pw()
+        else:
+            start()
     else:
-        start()
+        home = str(input( empty_list[0] + mode.lower() + empty_list[1] + yes_no ))
+        if home.lower() == "y":
+            start()
+        else:
+            exit()
 
 
 # Remove a password from the encrypted passwords table based on it's index
@@ -175,40 +193,48 @@ def del_pw():
                        """)
     list_pw = sql_cursor.fetchall()
 
-    # Loop through the contents of list_pw in human-readable format
-    place = 1
-    for i in list_pw:
-        print(place, " | ", i[0], " | ", i[-3], " | ", i[-1])
-        place += 1
+    # Check if list_pw is empty to avoid app crash when input expects user to give index of record that exists
+    if list_pw != []:
+        # Loop through the contents of list_pw in human-readable format
+        place = 1
+        for i in list_pw:
+            print(place, " | ", i[0], " | ", i[-3], " | ", i[-1])
+            place += 1
 
-    index = int(int(input( line + select[0] + mode.lower() + select[1] )) - 1)
+        index = int(int(input( line + select[0] + mode.lower() + select[1] )) - 1)
 
-    # Before deleting the password we need to save the returned account and username for the delete statement
-    acct = str(list_pw[index][0])
-    usr = str(list_pw[index][1])
+        # Before deleting the password we need to save the returned account and username for the delete statement
+        acct = str(list_pw[index][0])
+        usr = str(list_pw[index][1])
 
-    # Check if the user wants to delete the chosen pw
-    sure = str(input(f"{line}Are you sure you want to delete the password for {list_pw[index][0]} ?\n{yes_no}"))
-    if sure.lower() == "y":
+        # Check if the user wants to delete the chosen pw
+        sure = str(input(f"{line}Are you sure you want to delete the password for {list_pw[index][0]} ?\n{yes_no}"))
+        if sure.lower() == "y":
+            # Success statement needs to slice first letter off mode
+            print(mode[:-1] + doing)
+            sql_cursor.execute(f"""
+                            DELETE FROM passwords 
+                            WHERE account = '{acct}' 
+                            AND username = '{usr}';
+                            """)
+            sql_connection.commit()
+        elif sure.lower() == "n":
+            start()
+
+        # Return to Start Menu or repeat
         # Success statement needs to slice first letter off mode
-        print(mode[:-1] + doing)
-        sql_cursor.execute(f"""
-                           DELETE FROM passwords 
-                           WHERE account = '{acct}' 
-                           AND username = '{usr}';
-                           """)
-        sql_connection.commit()
-    elif sure.lower() == "n":
-        start()
-
-    # Return to Start Menu or repeat
-    # Success statement needs to slice first letter off mode
-    # Other funcs incl edit, add, drecypted so deleteed is wrong
-    again = str(input( line + mode[:-1] + done[0] + list_pw[index][0] + done[1] + mode + another ))
-    if again.lower() == "y":
-        del_pw()
+        # Other funcs incl edit, add, drecypted so deleteed is wrong
+        again = str(input( line + mode[:-1] + done[0] + list_pw[index][0] + done[1] + mode + another ))
+        if again.lower() == "y":
+            del_pw()
+        else:
+            start()
     else:
-        start()
+        home = str(input( empty_list[0] + mode.lower() + empty_list[1] + yes_no ))
+        if home.lower() == "y":
+            start()
+        else:
+            exit()
 
 
 # Choose a password to display based on it's index in the encrypted passwords table
@@ -224,27 +250,40 @@ def show_pw():
                        """)
     list_pw = sql_cursor.fetchall()
 
-    # Loop through the contents of list_pw in human-readable format
-    place = 1
-    for i in list_pw:
-        print(place, " | ", i[0], " | ", i[-3], " | ", i[-1])
-        place += 1
+    # Check if list_pw is empty to avoid app crash when input expects user to give index of record that exists
+    if list_pw != []:
+        # Loop through the contents of list_pw in human-readable format
+        place = 1
+        for i in list_pw:
+            print(place, " | ", i[0], " | ", i[-3], " | ", i[-1])
+            place += 1
 
-    index = int(int(input( line + select[0] + mode.lower() + select[1] )) - 1)
+        index = int(int(input( line + select[0] + mode.lower() + select[1] )) - 1)
 
-    print(mode + doing)
-    # Decrypted password needs to be saved to its own variable
-    decoded_pw = fernet.decrypt(list_pw[index][-3])
-    # Remember to decode the new variable to convert from BITS datatype to STRING
-    # This removes the leading b value changing b'variable' to 'variable'
-    print(f"\n{decoded_pw.decode()}\n")
+        # Before decrypting the password we need to save the returned Encryption Key for that record
+        pw_key = list_pw[index][-2]
 
-    # Return to Start Menu or repeat
-    again = str(input( line + mode + done[0] + list_pw[index][0] + done[1] + mode + another ))
-    if again.lower() == "y":
-        show_pw()
+        print(mode + doing)
+        # Decrypted password needs to be saved to its own variable
+        # We use Fernet(pw_key) here instead of fernet variable to
+        # Decrypt with the relevant records Encryption Key
+        decoded_pw = Fernet(pw_key).decrypt(list_pw[index][-3])
+        # Remember to decode the new variable to convert from BITS datatype to STRING
+        # This removes the leading b value changing b'variable' to 'variable'
+        print(f"\n{decoded_pw.decode()}\n")
+
+        # Return to Start Menu or repeat
+        again = str(input( line + mode + done[0] + list_pw[index][0] + done[1] + mode + another ))
+        if again.lower() == "y":
+            show_pw()
+        else:
+            start()
     else:
-        start()
+        home = str(input( empty_list[0] + mode.lower() + empty_list[1] + yes_no ))
+        if home.lower() == "y":
+            start()
+        else:
+            exit()
 
 
 # Start the CLI app
